requirements.txt
opencv_python==4.11.0.86
Pillow==11.2.1
torch==2.7.0
ultralytics==8.3.146
